Mavenize
- Make every lesson a Maven module. This loosens coupling, so that problems in one module do not prevent using another.
- Put ANTLRv4 grammars in default locations, so that generated source code doesn't clutter the source code.
- Turn `example.cfg` into a Java resource. This makes `nl.han.asd.app.lexing_parsing.Main` insusceptible to IDE 'Run configuration' working directory changes.
- Remove detailed lesson numbering from package names.
