+++ examples/evaluation_checking/src/main/antlr4/nl/han/asd/app/evaluation_checking/Configuration.g4
