+++ examples/evaluation_checking/src/main/java/nl/han/asd/app/evaluation_checking/ConfigurationReader.java
