+++ examples/generics/src/main/java/nl/han/asd/app/generics/MemoryCell.java
