+++ examples/lexing_parsing/src/main/antlr4/nl/han/asd/app/lexing_parsing/Configuration.g4
