+++ examples/lexing_parsing/src/main/resources/example.cfg
